JP-BE-PY-batch-2/class_24/CBV_extended/requirements.txt
djangorestframework
django-filter
drf-yasg